requirements.txt
flask
requests
python-dotenv
openai-whisper  # optional: only needed for voice notes
